Create Bit Stuffing
We are not using ASCII characters here. Any arbitrary number of bits is taken from the network layer. This stream of bits is taken and to it, it gives a starting and ending delimiter, which is the flag byte. Both are the same. But the problem is when the data itself is the same as the flag byte sequence. Then will be mistaken as a delimiter. 

To avoid these circumstances:

For every occurrence of 5 consecutive 1s, then the method of bit stuffing automatically stuffs a 0 after it. This happens on the sender side. After this process for all the data bits is over, the ending delimiter is attached. 

When the receiver sees a 0 after 5 consecutive 1s, it understands that it is a stuffed zero and then removes it. That hyphen indicates that in the example.